+++ contracts/core/l2/SubStorageL2.sol
@@ -27,20 +27,14 @@ contract SubStorageL2 is StrategyModel, AdminAuth, CoreHelper {
 
     /// @dev push one empty sub for AaveSubProxy to function correctly
     constructor() {
-        strategiesSubs.push(StoredSubData({
-            walletAddr: bytes20(0),
-            isEnabled: false,
-            strategySubHash: bytes32(0)
-        }));
+        strategiesSubs.push(StoredSubData({ walletAddr: bytes20(0), isEnabled: false, strategySubHash: bytes32(0) }));
 
-        storedStrategies.push(StrategySub({
-            strategyOrBundleId: 0,
-            isBundle: false,
-            triggerData: new bytes[](0),
-            subData: new bytes32[](0)
-        }));
+        storedStrategies.push(
+            StrategySub({
+                strategyOrBundleId: 0, isBundle: false, triggerData: new bytes[](0), subData: new bytes32[](0)
+            })
+        );
     }
-
 
     /// @notice Checks if subId is init. and if the sender is the owner
     modifier onlySubOwner(uint256 _subId) {
@@ -68,17 +62,14 @@ contract SubStorageL2 is StrategyModel, AdminAuth, CoreHelper {
     /// @notice Adds users info and records StoredSubData, logs StrategySub
     /// @dev To save on gas we don't store the whole struct but rather the hash of the struct
     /// @param _sub Subscription struct of the user (is not stored on chain, only the hash)
-    function subscribeToStrategy(
-        StrategySub memory _sub
-    ) public isValidId(_sub.strategyOrBundleId, _sub.isBundle) returns (uint256) {
-
+    function subscribeToStrategy(StrategySub memory _sub)
+        public
+        isValidId(_sub.strategyOrBundleId, _sub.isBundle)
+        returns (uint256)
+    {
         bytes32 subStorageHash = keccak256(abi.encode(_sub));
 
-        strategiesSubs.push(StoredSubData(
-            bytes20(msg.sender),
-            true,
-            subStorageHash
-        ));
+        strategiesSubs.push(StoredSubData(bytes20(msg.sender), true, subStorageHash));
 
         storedStrategies.push(_sub);
 
@@ -93,10 +84,11 @@ contract SubStorageL2 is StrategyModel, AdminAuth, CoreHelper {
     /// @dev Only callable by wallet who created the sub.
     /// @param _subId Id of the subscription to update
     /// @param _sub Subscription struct of the user (needs whole struct so we can hash it)
-    function updateSubData(
-        uint256 _subId,
-        StrategySub calldata _sub
-    ) public onlySubOwner(_subId) isValidId(_sub.strategyOrBundleId, _sub.isBundle)  {
+    function updateSubData(uint256 _subId, StrategySub calldata _sub)
+        public
+        onlySubOwner(_subId)
+        isValidId(_sub.strategyOrBundleId, _sub.isBundle)
+    {
         StoredSubData storage storedSubData = strategiesSubs[_subId];
 
         bytes32 subStorageHash = keccak256(abi.encode(_sub));
@@ -111,9 +103,7 @@ contract SubStorageL2 is StrategyModel, AdminAuth, CoreHelper {
     /// @notice Enables the subscription for execution if disabled
     /// @dev Must own the sub. to be able to enable it
     /// @param _subId Id of subscription to enable
-    function activateSub(
-        uint _subId
-    ) public onlySubOwner(_subId) {
+    function activateSub(uint256 _subId) public onlySubOwner(_subId) {
         StoredSubData storage sub = strategiesSubs[_subId];
 
         sub.isEnabled = true;
@@ -124,9 +114,7 @@ contract SubStorageL2 is StrategyModel, AdminAuth, CoreHelper {
     /// @notice Disables the subscription (will not be able to execute the strategy for the user)
     /// @dev Must own the sub. to be able to disable it
     /// @param _subId Id of subscription to disable
-    function deactivateSub(
-        uint _subId
-    ) public onlySubOwner(_subId) {
+    function deactivateSub(uint256 _subId) public onlySubOwner(_subId) {
         StoredSubData storage sub = strategiesSubs[_subId];
 
         sub.isEnabled = false;
@@ -136,11 +124,11 @@ contract SubStorageL2 is StrategyModel, AdminAuth, CoreHelper {
 
     ///////////////////// VIEW ONLY FUNCTIONS ////////////////////////////
 
-    function getSub(uint _subId) public view returns (StoredSubData memory) {
+    function getSub(uint256 _subId) public view returns (StoredSubData memory) {
         return strategiesSubs[_subId];
     }
 
-    function getStrategySub(uint _subId) public view returns (StrategySub memory) {
+    function getStrategySub(uint256 _subId) public view returns (StrategySub memory) {
         return storedStrategies[_subId];
     }
