+++ contracts/interfaces/liquity/IPriceFeed.sol
@@ -4,5 +4,5 @@ pragma solidity =0.8.24;
 
 interface IPriceFeed {
     function lastGoodPrice() external pure returns (uint256);
-    function fetchPrice() external returns (uint);
+    function fetchPrice() external returns (uint256);
 }
