+++ contracts/utils/math/Exponential.sol
@@ -7,17 +7,17 @@ import { CarefulMath } from "./CarefulMath.sol";
 /// @title Exponential module for storing fixed-precision decimals
 /// @author Compound
 contract Exponential is CarefulMath {
-    uint constant expScale = 1e18;
-    uint constant doubleScale = 1e36;
-    uint constant halfExpScale = expScale/2;
-    uint constant mantissaOne = expScale;
+    uint256 constant expScale = 1e18;
+    uint256 constant doubleScale = 1e36;
+    uint256 constant halfExpScale = expScale / 2;
+    uint256 constant mantissaOne = expScale;
 
     struct Exp {
-        uint mantissa;
+        uint256 mantissa;
     }
 
     struct Double {
-        uint mantissa;
+        uint256 mantissa;
     }
 
     /**
@@ -25,54 +25,54 @@ contract Exponential is CarefulMath {
      *      Note: Returns an error if (`num` * 10e18) > MAX_INT,
      *            or if `denom` is zero.
      */
-    function getExp(uint num, uint denom) internal pure returns (MathError, Exp memory) {
-        (MathError err0, uint scaledNumerator) = mulUInt(num, expScale);
+    function getExp(uint256 num, uint256 denom) internal pure returns (MathError, Exp memory) {
+        (MathError err0, uint256 scaledNumerator) = mulUInt(num, expScale);
         if (err0 != MathError.NO_ERROR) {
-            return (err0, Exp({mantissa: 0}));
+            return (err0, Exp({ mantissa: 0 }));
         }
 
-        (MathError err1, uint rational) = divUInt(scaledNumerator, denom);
+        (MathError err1, uint256 rational) = divUInt(scaledNumerator, denom);
         if (err1 != MathError.NO_ERROR) {
-            return (err1, Exp({mantissa: 0}));
+            return (err1, Exp({ mantissa: 0 }));
         }
 
-        return (MathError.NO_ERROR, Exp({mantissa: rational}));
+        return (MathError.NO_ERROR, Exp({ mantissa: rational }));
     }
 
     /**
      * @dev Adds two exponentials, returning a new exponential.
      */
     function addExp(Exp memory a, Exp memory b) internal pure returns (MathError, Exp memory) {
-        (MathError error, uint result) = addUInt(a.mantissa, b.mantissa);
+        (MathError error, uint256 result) = addUInt(a.mantissa, b.mantissa);
 
-        return (error, Exp({mantissa: result}));
+        return (error, Exp({ mantissa: result }));
     }
 
     /**
      * @dev Subtracts two exponentials, returning a new exponential.
      */
     function subExp(Exp memory a, Exp memory b) internal pure returns (MathError, Exp memory) {
-        (MathError error, uint result) = subUInt(a.mantissa, b.mantissa);
+        (MathError error, uint256 result) = subUInt(a.mantissa, b.mantissa);
 
-        return (error, Exp({mantissa: result}));
+        return (error, Exp({ mantissa: result }));
     }
 
     /**
      * @dev Multiply an Exp by a scalar, returning a new Exp.
      */
-    function mulScalar(Exp memory a, uint scalar) internal pure returns (MathError, Exp memory) {
-        (MathError err0, uint scaledMantissa) = mulUInt(a.mantissa, scalar);
+    function mulScalar(Exp memory a, uint256 scalar) internal pure returns (MathError, Exp memory) {
+        (MathError err0, uint256 scaledMantissa) = mulUInt(a.mantissa, scalar);
         if (err0 != MathError.NO_ERROR) {
-            return (err0, Exp({mantissa: 0}));
+            return (err0, Exp({ mantissa: 0 }));
         }
 
-        return (MathError.NO_ERROR, Exp({mantissa: scaledMantissa}));
+        return (MathError.NO_ERROR, Exp({ mantissa: scaledMantissa }));
     }
 
     /**
      * @dev Multiply an Exp by a scalar, then truncate to return an unsigned integer.
      */
-    function mulScalarTruncate(Exp memory a, uint scalar) internal pure returns (MathError, uint) {
+    function mulScalarTruncate(Exp memory a, uint256 scalar) internal pure returns (MathError, uint256) {
         (MathError err, Exp memory product) = mulScalar(a, scalar);
         if (err != MathError.NO_ERROR) {
             return (err, 0);
@@ -84,7 +84,11 @@ contract Exponential is CarefulMath {
     /**
      * @dev Multiply an Exp by a scalar, truncate, then add an to an unsigned integer, returning an unsigned integer.
      */
-    function mulScalarTruncateAddUInt(Exp memory a, uint scalar, uint addend) internal pure returns (MathError, uint) {
+    function mulScalarTruncateAddUInt(Exp memory a, uint256 scalar, uint256 addend)
+        internal
+        pure
+        returns (MathError, uint256)
+    {
         (MathError err, Exp memory product) = mulScalar(a, scalar);
         if (err != MathError.NO_ERROR) {
             return (err, 0);
@@ -96,19 +100,19 @@ contract Exponential is CarefulMath {
     /**
      * @dev Divide an Exp by a scalar, returning a new Exp.
      */
-    function divScalar(Exp memory a, uint scalar) internal pure returns (MathError, Exp memory) {
-        (MathError err0, uint descaledMantissa) = divUInt(a.mantissa, scalar);
+    function divScalar(Exp memory a, uint256 scalar) internal pure returns (MathError, Exp memory) {
+        (MathError err0, uint256 descaledMantissa) = divUInt(a.mantissa, scalar);
         if (err0 != MathError.NO_ERROR) {
-            return (err0, Exp({mantissa: 0}));
+            return (err0, Exp({ mantissa: 0 }));
         }
 
-        return (MathError.NO_ERROR, Exp({mantissa: descaledMantissa}));
+        return (MathError.NO_ERROR, Exp({ mantissa: descaledMantissa }));
     }
 
     /**
      * @dev Divide a scalar by an Exp, returning a new Exp.
      */
-    function divScalarByExp(uint scalar, Exp memory divisor) internal pure returns (MathError, Exp memory) {
+    function divScalarByExp(uint256 scalar, Exp memory divisor) internal pure returns (MathError, Exp memory) {
         /*
           We are doing this as:
           getExp(mulUInt(expScale, scalar), divisor.mantissa)
@@ -118,9 +122,9 @@ contract Exponential is CarefulMath {
           Scalar = s;
           `s / (a / b)` = `b * s / a` and since for an Exp `a = mantissa, b = expScale`
         */
-        (MathError err0, uint numerator) = mulUInt(expScale, scalar);
+        (MathError err0, uint256 numerator) = mulUInt(expScale, scalar);
         if (err0 != MathError.NO_ERROR) {
-            return (err0, Exp({mantissa: 0}));
+            return (err0, Exp({ mantissa: 0 }));
         }
         return getExp(numerator, divisor.mantissa);
     }
@@ -128,7 +132,7 @@ contract Exponential is CarefulMath {
     /**
      * @dev Divide a scalar by an Exp, then truncate to return an unsigned integer.
      */
-    function divScalarByExpTruncate(uint scalar, Exp memory divisor) internal pure returns (MathError, uint) {
+    function divScalarByExpTruncate(uint256 scalar, Exp memory divisor) internal pure returns (MathError, uint256) {
         (MathError err, Exp memory fraction) = divScalarByExp(scalar, divisor);
         if (err != MathError.NO_ERROR) {
             return (err, 0);
@@ -141,32 +145,31 @@ contract Exponential is CarefulMath {
      * @dev Multiplies two exponentials, returning a new exponential.
      */
     function mulExp(Exp memory a, Exp memory b) internal pure returns (MathError, Exp memory) {
-
-        (MathError err0, uint doubleScaledProduct) = mulUInt(a.mantissa, b.mantissa);
+        (MathError err0, uint256 doubleScaledProduct) = mulUInt(a.mantissa, b.mantissa);
         if (err0 != MathError.NO_ERROR) {
-            return (err0, Exp({mantissa: 0}));
+            return (err0, Exp({ mantissa: 0 }));
         }
 
         // We add half the scale before dividing so that we get rounding instead of truncation.
         //  See "Listing 6" and text above it at https://accu.org/index.php/journals/1717
         // Without this change, a result like 6.6...e-19 will be truncated to 0 instead of being rounded to 1e-18.
-        (MathError err1, uint doubleScaledProductWithHalfScale) = addUInt(halfExpScale, doubleScaledProduct);
+        (MathError err1, uint256 doubleScaledProductWithHalfScale) = addUInt(halfExpScale, doubleScaledProduct);
         if (err1 != MathError.NO_ERROR) {
-            return (err1, Exp({mantissa: 0}));
+            return (err1, Exp({ mantissa: 0 }));
         }
 
-        (MathError err2, uint product) = divUInt(doubleScaledProductWithHalfScale, expScale);
+        (MathError err2, uint256 product) = divUInt(doubleScaledProductWithHalfScale, expScale);
         // The only error `div` can return is MathError.DIVISION_BY_ZERO but we control `expScale` and it is not zero.
         assert(err2 == MathError.NO_ERROR);
 
-        return (MathError.NO_ERROR, Exp({mantissa: product}));
+        return (MathError.NO_ERROR, Exp({ mantissa: product }));
     }
 
     /**
      * @dev Multiplies two exponentials given their mantissas, returning a new exponential.
      */
-    function mulExp(uint a, uint b) internal pure returns (MathError, Exp memory) {
-        return mulExp(Exp({mantissa: a}), Exp({mantissa: b}));
+    function mulExp(uint256 a, uint256 b) internal pure returns (MathError, Exp memory) {
+        return mulExp(Exp({ mantissa: a }), Exp({ mantissa: b }));
     }
 
     /**
@@ -193,7 +196,7 @@ contract Exponential is CarefulMath {
      * @dev Truncates the given exp to a whole number value.
      *      For example, truncate(Exp{mantissa: 15 * expScale}) = 15
      */
-    function truncate(Exp memory exp) internal pure returns (uint) {
+    function truncate(Exp memory exp) internal pure returns (uint256) {
         // Note: We are not using careful math here as we're performing a division that cannot fail
         return exp.mantissa / expScale;
     }
@@ -227,106 +230,106 @@ contract Exponential is CarefulMath {
     }
 
     function sub_(Exp memory a, Exp memory b) internal pure returns (Exp memory) {
-        return Exp({mantissa: sub_(a.mantissa, b.mantissa)});
+        return Exp({ mantissa: sub_(a.mantissa, b.mantissa) });
     }
 
     function sub_(Double memory a, Double memory b) internal pure returns (Double memory) {
-        return Double({mantissa: sub_(a.mantissa, b.mantissa)});
+        return Double({ mantissa: sub_(a.mantissa, b.mantissa) });
     }
 
-    function sub_(uint a, uint b) internal pure returns (uint) {
+    function sub_(uint256 a, uint256 b) internal pure returns (uint256) {
         return sub_(a, b, "subtraction underflow");
     }
 
-    function sub_(uint a, uint b, string memory errorMessage) internal pure returns (uint) {
+    function sub_(uint256 a, uint256 b, string memory errorMessage) internal pure returns (uint256) {
         require(b <= a, errorMessage);
         return a - b;
     }
 
     function mul_(Exp memory a, Exp memory b) internal pure returns (Exp memory) {
-        return Exp({mantissa: mul_(a.mantissa, b.mantissa) / expScale});
+        return Exp({ mantissa: mul_(a.mantissa, b.mantissa) / expScale });
     }
 
-    function mul_(Exp memory a, uint b) internal pure returns (Exp memory) {
-        return Exp({mantissa: mul_(a.mantissa, b)});
+    function mul_(Exp memory a, uint256 b) internal pure returns (Exp memory) {
+        return Exp({ mantissa: mul_(a.mantissa, b) });
     }
 
-    function mul_(uint a, Exp memory b) internal pure returns (uint) {
+    function mul_(uint256 a, Exp memory b) internal pure returns (uint256) {
         return mul_(a, b.mantissa) / expScale;
     }
 
     function mul_(Double memory a, Double memory b) internal pure returns (Double memory) {
-        return Double({mantissa: mul_(a.mantissa, b.mantissa) / doubleScale});
+        return Double({ mantissa: mul_(a.mantissa, b.mantissa) / doubleScale });
     }
 
-    function mul_(Double memory a, uint b) internal pure returns (Double memory) {
-        return Double({mantissa: mul_(a.mantissa, b)});
+    function mul_(Double memory a, uint256 b) internal pure returns (Double memory) {
+        return Double({ mantissa: mul_(a.mantissa, b) });
     }
 
-    function mul_(uint a, Double memory b) internal pure returns (uint) {
+    function mul_(uint256 a, Double memory b) internal pure returns (uint256) {
         return mul_(a, b.mantissa) / doubleScale;
     }
 
-    function mul_(uint a, uint b) internal pure returns (uint) {
+    function mul_(uint256 a, uint256 b) internal pure returns (uint256) {
         return mul_(a, b, "multiplication overflow");
     }
 
-    function mul_(uint a, uint b, string memory errorMessage) internal pure returns (uint) {
+    function mul_(uint256 a, uint256 b, string memory errorMessage) internal pure returns (uint256) {
         if (a == 0 || b == 0) {
             return 0;
         }
-        uint c = a * b;
+        uint256 c = a * b;
         require(c / a == b, errorMessage);
         return c;
     }
 
     function div_(Exp memory a, Exp memory b) internal pure returns (Exp memory) {
-        return Exp({mantissa: div_(mul_(a.mantissa, expScale), b.mantissa)});
+        return Exp({ mantissa: div_(mul_(a.mantissa, expScale), b.mantissa) });
     }
 
-    function div_(Exp memory a, uint b) internal pure returns (Exp memory) {
-        return Exp({mantissa: div_(a.mantissa, b)});
+    function div_(Exp memory a, uint256 b) internal pure returns (Exp memory) {
+        return Exp({ mantissa: div_(a.mantissa, b) });
     }
 
-    function div_(uint a, Exp memory b) internal pure returns (uint) {
+    function div_(uint256 a, Exp memory b) internal pure returns (uint256) {
         return div_(mul_(a, expScale), b.mantissa);
     }
 
     function div_(Double memory a, Double memory b) internal pure returns (Double memory) {
-        return Double({mantissa: div_(mul_(a.mantissa, doubleScale), b.mantissa)});
+        return Double({ mantissa: div_(mul_(a.mantissa, doubleScale), b.mantissa) });
     }
 
-    function div_(Double memory a, uint b) internal pure returns (Double memory) {
-        return Double({mantissa: div_(a.mantissa, b)});
+    function div_(Double memory a, uint256 b) internal pure returns (Double memory) {
+        return Double({ mantissa: div_(a.mantissa, b) });
     }
 
-    function div_(uint a, Double memory b) internal pure returns (uint) {
+    function div_(uint256 a, Double memory b) internal pure returns (uint256) {
         return div_(mul_(a, doubleScale), b.mantissa);
     }
 
-    function div_(uint a, uint b) internal pure returns (uint) {
+    function div_(uint256 a, uint256 b) internal pure returns (uint256) {
         return div_(a, b, "divide by zero");
     }
 
-    function div_(uint a, uint b, string memory errorMessage) internal pure returns (uint) {
+    function div_(uint256 a, uint256 b, string memory errorMessage) internal pure returns (uint256) {
         require(b > 0, errorMessage);
         return a / b;
     }
 
     function add_(Exp memory a, Exp memory b) internal pure returns (Exp memory) {
-        return Exp({mantissa: add_(a.mantissa, b.mantissa)});
+        return Exp({ mantissa: add_(a.mantissa, b.mantissa) });
     }
 
     function add_(Double memory a, Double memory b) internal pure returns (Double memory) {
-        return Double({mantissa: add_(a.mantissa, b.mantissa)});
+        return Double({ mantissa: add_(a.mantissa, b.mantissa) });
     }
 
-    function add_(uint a, uint b) internal pure returns (uint) {
+    function add_(uint256 a, uint256 b) internal pure returns (uint256) {
         return add_(a, b, "addition overflow");
     }
 
-    function add_(uint a, uint b, string memory errorMessage) internal pure returns (uint) {
-        uint c = a + b;
+    function add_(uint256 a, uint256 b, string memory errorMessage) internal pure returns (uint256) {
+        uint256 c = a + b;
         require(c >= a, errorMessage);
         return c;
     }
